% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportCoverage}
\alias{exportCoverage}
\title{Export coverage as BigWig track}
\usage{
exportCoverage(coverage, filename = 'wavClusters.BigWig')
}
\arguments{
\item{coverage}{An Rle object containing the coverage at each genomic
position as returned by a call to \code{coverage}}

\item{filename}{A character defining the BED file name. Default to
"wavClusters.BigWig"}
}
\value{
A BigWig file of the exported Rle object
}
\description{
Export coverage as BigWig track, compatible with the UCSC genome browser
}
\author{
Federico Comoglio
}
\keyword{postprocessing}

